+++ browser/portal/portal.menu.item.js
@@ -1,23 +1,28 @@
 // @ts-check
 
-/** @type { import("../..").handlerFactory } */
+/** @type { import("../../handlers").handlerFactory } */
 module.exports = ({
     config: {
         pages
     } = {}
 }) => ({
-    async 'portal.menu.item'(params) {
-        const {id, title, ...rest} = params;
-        if ('component' in rest) delete rest.component;
+    async 'portal.menu.item'(params, $meta) {
+        let {id, title, component: _, ...rest} = params;
         const component = typeof params === 'function' ? params : params.component;
         let page;
         try {
             page = await component({});
         } catch (e) {
+            if (title) title = '💥' + title;
             page = {
-                title: component.name,
+                title: '💥' + component.name,
+                component: null,
                 disabled: true
             };
+            // eslint-disable-next-line no-process-env
+            if (process.env.NODE_ENV !== 'production') {
+                this.log?.warn?.(`Menu item ${component.name} returned error ${e.message}`);
+            }
         }
         const name = component.name.split('/').pop();
         let query = '';