+++ index.js
@@ -1,4 +1,9 @@
 module.exports = () => function utPortal() {
     return {
+        config: require('./config'),
+        test: () => [
+            ...require('./test/steps'),
+            ...require('./test/jobs')
+        ]
     };
 };