+++ test/steps/playwright.config.js
@@ -1,0 +1,31 @@
+/* eslint-disable no-process-env */
+// @ts-check
+
+const {resolve} = require('path');
+
+/** @type {import('@playwright/test').PlaywrightTestConfig<{ username: string, password: string }>} */
+const config = {
+    retries: 1,
+    use: {
+        trace: 'on-first-retry',
+        ...process.env.CHROME_BIN && {
+            launchOptions: {
+                executablePath: process.env.CHROME_BIN
+            }
+        },
+        baseURL: process.env.UT_URL,
+        viewport: { width: 1600, height: 900 },
+        username: process.env.UT_USERNAME,
+        password: process.env.UT_PASSWORD
+    },
+    testDir: resolve('.').replaceAll('\\', '/'),
+    testMatch: /.*\.play\.(js|ts|mjs)/,
+    outputDir: resolve('.lint/playwright'),
+    reporter: [
+        [process.env.CI ? 'dot' : 'list'],
+        ['html', { open: 'never', outputFolder: resolve('.lint/playwright-unit') }],
+        ['junit', { outputFile: resolve('.lint/playwright-unit.xml') }]
+    ]
+};
+
+module.exports = config;